+++ index.php
@@ -17,13 +17,15 @@
 	
 	($_SERVER['SCRIPT_NAME'] !== "/controller.php") ? require_once(__DIR__ . "/classes/Requires.php") : Links::$pretty = true;
 	
+	$db_version = Functions::get_db_version();
+	
 	Header::set_title("Commando.io");
 	Header::render();
 	
 	Navigation::render();
 ?>
 
- 	<?php if(Functions::get_db_version() !== Version::db): ?>
+ 	<?php if($db_version !== Version::db): ?>
 		<div class="alert alert-error fade in" style="position: fixed; top: 0px; right: 0px; left: 0px; z-index: 99999;">
 			<a class="close" data-dismiss="alert">&times;</a>
 			<h4 style="line-height: 25px;">Warning! The MySQL schema version you are running is <u>out of date</u>. Please run <a href="/db-upgrade.php">db-upgrade.php</a> to merge in the latest changes.</h4>